Test snap set done, save before change snap prefabs
Test prefabs in scene are snapping correctly with my preautomation. Added a "ConnectionPose" to have AttachPoint and SnapPoint aligned with it in its positioning as right where atom and bond connect. Last save before updating the snapping prefabs. Come back to this if it messes all the snapping up. Still having an "unassigned reference exception" related to the SnapToPosition script.